Move grdview function to grdview.py
